GRID | added basic comments to grid component
